% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cell_fire_status}
\alias{cell_fire_status}
\title{Determine fire status given interval thresholds and fire history}
\usage{
cell_fire_status(
  fireyears,
  min_threshold,
  max_threshold,
  year_query,
  year_base,
  quiet = TRUE
)
}
\arguments{
\item{fireyears}{Vector of integer fire years. Any years beyond year_query
are ignored, as are duplicate years.}

\item{min_threshold}{Minimum tolerable fire interval (years).}

\item{max_threshold}{Maximum tolerable fire interval (years).}

\item{year_query}{Reference year for the analysis.}

\item{year_base}{Year taken as the start of the data period.}

\item{quiet}{If true (default) do not print debugging output.}
}
\value{
An integer code for fire status.
}
\description{
This function determines the fire status of a location (e.g. point or
raster cell) given a vector of fire years, and the minimum and maximum
tolerable fire intervals (years) for the vegetation at the location.
}
