% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_status.R
\name{get_fire_status}
\alias{get_fire_status}
\title{Fire interval threshold analysis for one or more point locations or raster cells}
\usage{
get_fire_status(firehistory, veg, thresholds, query_years, base_year)
}
\arguments{
\item{firehistory}{A matrix or data frame with two columns: integer
location ID (e.g. raster cell number); and integer fire year.}

\item{veg}{A matrix or data frame with two columns: integer location ID
(e.g. raster cell number); and integer vegetation type code. There must
only be one row for each location ID.}

\item{thresholds}{A matrix or data frame with three columns: integer veg
code; minimum tolerable interval (years); maximum tolerable interval
(years). Setting both the minimum and maximum values to zero indicates that
the vegetation type has no fire regime defined. Setting both values to
\code{NA} indicates that the vegetation type should never be burnt.}

\item{query_years}{One or more reference years for which to report fire
status.}

\item{base_year}{The earliest year covered by the provided fire history. This
can be earlier than any actual fire years specified for locations.}
}
\description{
Fire interval threshold analysis for one or more point locations or raster cells
}
