% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{table_fire_status}
\alias{table_fire_status}
\title{Determine fire status for a set of locations}
\usage{
table_fire_status(
  cell_firehistory,
  cell_veg,
  veg_thresholds,
  year_query,
  year_base,
  quiet = TRUE
)
}
\description{
Determine fire status for a set of locations
}
