% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_status.R
\name{get_fire_status}
\alias{get_fire_status}
\title{Fire interval threshold analysis for one or more point locations or raster cells}
\usage{
get_fire_status(
  firehistory,
  veg,
  thresholds,
  query_years,
  base_year,
  estimate_base_tsf
)
}
\arguments{
\item{firehistory}{A matrix or data frame with two columns: integer
location ID (e.g. raster cell number); and integer fire year.}

\item{veg}{A matrix or data frame with two columns: integer location ID
(e.g. raster cell number); and integer vegetation type code. There must
only be one row for each location ID.}

\item{thresholds}{A matrix or data frame with three columns: integer veg
code; minimum tolerable interval (years); maximum tolerable interval
(years). Setting both the minimum and maximum values to zero indicates that
the vegetation type has no fire regime defined. Setting both values to
\code{NA} indicates that the vegetation type should never be burnt.}

\item{query_years}{One or more reference years for which to report fire
status.}

\item{base_year}{Default start year of the data period. If no fires have
occurred, and \code{estimate_base_tsf == TRUE} a time since fire value for
the query year is set by assuming that a fire occurred N years before the
base_year, where N is the mid-point of the vegetation's tolerable fire
interval. If fires have occurred, but only after the base year, then we
assume an earlier fire based on the same mid-point calculation. This only
applies if the minimum and maximum threshold values are not 0 (no fire
regime) or 9999 (vegetation that should not be burnt.)}

\item{estimate_base_tsf}{If \code{TRUE} and no fire years are earlier than
\code{base_year}, a pre-base year fire date is assumed based on the
mid-point of interval thresholds (see above). This only applies if
the minimum and maximum threshold values are not 0 (no fire regime) or
9999 (vegetation that should not be burnt.)}
}
\value{
A matrix where the first column is location ID, and subsequent
  columns are fire status for each query year.
}
\description{
Fire interval threshold analysis for one or more point locations or raster cells
}
